src/test/java/br/com/hackerbrasileiro/webapp/domain/HackerTest.java
/*
Copyright (C) 2015  ThoughtWorks, Inc.

This program is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

package br.com.hackerbrasileiro.webapp.domain;


import org.junit.Test;

import static org.hamcrest.CoreMatchers.is;
import static org.hamcrest.MatcherAssert.assertThat;

public class HackerTest {

    @Test
    public void buildWithTest() {
        String line = "firstName,lastName,email";
        Hacker hacker = Hacker.buildWith(line);

        assertThat(hacker.getFirstName(), is("firstName"));
        assertThat(hacker.getLastName(), is("lastName"));
        assertThat(hacker.getEmail(), is("email"));
    }

}
